[기능추가] 간단한 저축챌린지를 위한 기능 추가
